requirements.txt
# Proxies may be necessary to run installation successfully
--index-url http://pyats-pypi.cisco.com
--trusted-host pyats-pypi.cisco.com
pyats
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
PyYAML==6.0
sqlparse==0.4.2
Django
Djangorestframework
celery
redis
